fix sorting bug. add check for exceptions during evaluating diagnostics
