+++ plugins/org.eclipse.n4js.ide/src/org/eclipse/n4js/ide/server/N4JSLanguageServerImpl.java
@@ -11,7 +11,6 @@
 package org.eclipse.n4js.ide.server;
 
 import java.util.Iterator;
-import java.util.concurrent.CompletableFuture;
 
 import org.eclipse.emf.common.util.URI;
 import org.eclipse.lsp4j.DidChangeTextDocumentParams;
@@ -19,7 +18,6 @@ import org.eclipse.lsp4j.DidChangeWatchedFilesParams;
 import org.eclipse.lsp4j.DidCloseTextDocumentParams;
 import org.eclipse.lsp4j.DidOpenTextDocumentParams;
 import org.eclipse.lsp4j.FileEvent;
-import org.eclipse.lsp4j.InitializedParams;
 import org.eclipse.lsp4j.services.LanguageClient;
 import org.eclipse.n4js.hlc.base.HeadlessExtensionRegistrationHelper;
 import org.eclipse.n4js.ide.xtext.server.XLanguageServerImpl;
@@ -46,21 +44,6 @@ public class N4JSLanguageServerImpl extends XLanguageServerImpl {
 	public void registerExtensions(HeadlessExtensionRegistrationHelper helper) {
 		helper.unregisterExtensions();
 		helper.registerExtensions();
-	}
-
-	/**
-	 * Call this method after calling {@link #initialized(InitializedParams)}
-	 *
-	 * TODO (SZ): initialized(params) calls CompleteableFuture.complete - the join is necessary to see exceptions or
-	 * what is the purpose?
-	 *
-	 * <p>
-	 * TODO: Fix this in Xtext
-	 */
-	public void joinInitialized() {
-		CompletableFuture<InitializedParams> initialized = ReflectionUtils.getFieldValue(XLanguageServerImpl.class,
-				this, "initialized");
-		initialized.join();
 	}
 
 	@Override
